requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
jsonpickle==4.0.1
python-digitalocean==1.17.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
